Split the fixed tests into multiple classes
