+++ Exanite.Core.Tests/Numerics/FixedTestConstants.cs
@@ -1,0 +1,12 @@
+using Exanite.Core.Numerics;
+
+namespace Exanite.Core.Tests.Numerics;
+
+public static class FixedTestConstants
+{
+    /// <summary>
+    /// The default decimal places of precision that is
+    /// expected for the <see cref="Fixed"/> data type.
+    /// </summary>
+    public static int BaseExpectedPrecision => 4;
+}
